Move truthy to utils component
