+++ app/app.js
@@ -17,7 +17,6 @@ const v2Router = require('./src/routes/v2');
 const { authorizedParty } = require('./src/middleware/authorizedParty');
 const initializeApiTracker = require('./src/middleware/apiTracker');
 
-const carboneCopyApi = require('./src/components/carboneCopyApi');
 const carboneCopyMiddleware = require('./src/middleware/carboneCopy');
 const carboneBasePath = '/api/v3';
 
@@ -123,7 +122,7 @@ app.get('/api-spec.yaml', (req, res) => {
   req.url = `${carboneBasePath}/api-spec.yaml`;
   app.handle(req, res);
 });
-app.use(carboneBasePath, carboneCopyMiddleware.operation(carboneBasePath), carboneCopyMiddleware.security, carboneCopyMiddleware.cacheCleanup, carboneCopyApi.routes());
+app.use(carboneBasePath, carboneCopyMiddleware.operation(carboneBasePath), carboneCopyMiddleware.security, carboneCopyMiddleware.cacheCleanup);
 
 // Root level Router
 app.use(/(\/api)?/, apiRouter);
