+++ app/src/components/carboneCopyApi.js
@@ -1,5 +1,3 @@
-const router = require('express').Router();
-
 const mime = require('mime-types');
 const path = require('path');
 const Problem = require('api-problem');
@@ -8,6 +6,7 @@ const telejson = require('telejson');
 const carboneRenderer = require('./carboneRender');
 const FileCache = require('./fileCache');
 const fileUpload = require('./upload');
+const { truthy } = require('./utils');
 
 const CACHE_DIR = process.env.CACHE_DIR || '/tmp/carbone-files';
 const CONVERTER_FACTORY_TIMEOUT = process.env.CONVERTER_FACTORY_TIMEOUT || '60000';
@@ -27,11 +26,6 @@ const DEFAULT_OPTIONS = {
 
 let fileCache;
 
-const truthy = (name, obj = {}) => {
-  const value = obj[name] || false;
-  return (value === 'true' || value === '1' || value === 'yes' || value === 'y' || value === 't' || value === 1 || value === true);
-};
-
 const renderTemplate = async (template, req, res) => {
   let data = req.body.data;
   let options = {};
@@ -40,7 +34,7 @@ const renderTemplate = async (template, req, res) => {
   try {
     options = req.body.options;
   } catch (e) {
-    return new Problem(400, {detail: 'options not provided or formatted incorrectly'}).send(res);
+    return new Problem(400, { detail: 'options not provided or formatted incorrectly' }).send(res);
   }
 
   options.convertTo = options.convertTo || template.ext;
@@ -58,7 +52,7 @@ const renderTemplate = async (template, req, res) => {
     try {
       data = req.body.data;
     } catch (e) {
-      return new Problem(400, {detail: 'data not provided or formatted incorrectly'}).send(res);
+      return new Problem(400, { detail: 'data not provided or formatted incorrectly' }).send(res);
     }
   }
 
@@ -86,14 +80,14 @@ const renderTemplate = async (template, req, res) => {
 
     return res.send(output.report);
   } else {
-    return new Problem(output.errorType, {detail: output.errorMsg}).send(res);
+    return new Problem(output.errorType, { detail: output.errorMsg }).send(res);
   }
 };
 
 const getFromCache = async (hash, hashHeaderName, download, remove, res) => {
   const file = fileCache.find(hash);
   if (!file.success) {
-    return new Problem(file.errorType, {detail: file.errorMsg}).send(res);
+    return new Problem(file.errorType, { detail: file.errorMsg }).send(res);
   }
 
   let cached = undefined;
@@ -101,14 +95,14 @@ const getFromCache = async (hash, hashHeaderName, download, remove, res) => {
     try {
       cached = await fileCache.read(hash);
     } catch (e) {
-      return new Problem(500, {detail: e.message}).send(res);
+      return new Problem(500, { detail: e.message }).send(res);
     }
   }
 
   if (remove) {
     const removed = await fileCache.remove(hash);
     if (!removed.success) {
-      return new Problem(removed.errorType, {detail: removed.errorMsg}).send(res);
+      return new Problem(removed.errorType, { detail: removed.errorMsg }).send(res);
     }
   }
 
@@ -124,42 +118,31 @@ const getFromCache = async (hash, hashHeaderName, download, remove, res) => {
   }
 };
 
-const findAndRender = async (hash, req, res) => {
+const findAndRender = (hash, req, res) => {
   const template = fileCache.find(hash);
   if (!template.success) {
-    return new Problem(template.errorType, {detail: template.errorMsg}).send(res);
+    return new Problem(template.errorType, { detail: template.errorMsg }).send(res);
   } else {
-    return await renderTemplate(template, req, res);
+    return renderTemplate(template, req, res);
   }
 };
 
-let initialized = false;
 module.exports = {
   // Temporary Exports
   findAndRender: findAndRender,
   getFromCache: getFromCache,
-  truthy: truthy,
 
   init(options) {
-
     let _options = DEFAULT_OPTIONS;
     if (options) {
-      _options = {..._options, ...options};
+      _options = { ..._options, ...options };
     }
 
-    fileCache = new FileCache({fileCachePath: _options.fileUploadsDir});
+    fileCache = new FileCache({ fileCachePath: _options.fileUploadsDir });
     fileUpload.init();
 
     if (truthy('startCarbone', _options)) {
       carboneRenderer.startFactory(_options.converterFactoryTimeout);
     }
-    initialized = true;
-  },
-
-  routes(options) {
-    if (!initialized) {
-      this.init(options);
-    }
-    return router;
-  },
+  }
 };
